ycimage-main/requirements.txt
Pillow==10.2.0
wxPython==4.2.1
piexif==1.1.3
pywin32==306; platform_system == "Windows"
pillow_heif==0.15.0
matplotlib==3.8.3